+++ src/main/java/com/kamesuta/physxmc/core/PhysxTerrain.java
@@ -1,10 +1,6 @@
 package com.kamesuta.physxmc.core;
 
 import lombok.Getter;
-import org.bukkit.Chunk;
-import org.bukkit.World;
-import org.bukkit.block.Block;
-import org.bukkit.block.BlockFace;
 import physx.common.PxBase;
 import physx.common.PxIDENTITYEnum;
 import physx.common.PxTransform;
@@ -24,63 +20,43 @@ public class PhysxTerrain {
      * 地形本体
      */
     @Getter
-    private final PxRigidStatic actor;
+    protected PxRigidStatic actor;
 
     /**
      * 地形にアタッチされた形状たち
      */
-    private final List<PxShape> terrainShapes = new ArrayList<>();
+    protected final List<PxShape> terrainShapes = new ArrayList<>();
 
     public static final String name = "terrain";
 
     /**
-     * チャンクの形に応じた地形を作る
+     * y=-61にフラットな地形を作る
      *
      * @param defaultMaterial 　地形のマテリアル
-     * @param chunk           チャンク
      * @return 地形のオブジェクト
      */
-    public PhysxTerrain(PxPhysics physics, PxMaterial defaultMaterial, Chunk chunk) {
+    public PhysxTerrain(PxPhysics physics, PxMaterial defaultMaterial) {
+        if(physics == null)
+            return;
+        
         // create default simulation shape flags
         PxShapeFlags defaultShapeFlags = new PxShapeFlags((byte) (PxShapeFlagEnum.eSCENE_QUERY_SHAPE.value | PxShapeFlagEnum.eSIMULATION_SHAPE.value));
         // create a few temporary objects used during setup
         PxTransform tmpPose = new PxTransform(PxIDENTITYEnum.PxIdentity);
-        PxVec3 tmpVec = new PxVec3(chunk.getX() * 16, 0f, chunk.getZ() * 16);
-        tmpPose.setP(tmpVec);
-        tmpVec.destroy();
         PxFilterData tmpFilterData = new PxFilterData(1, 1, 0, 0);
-        PxRigidStatic terrain = physics.createRigidStatic(tmpPose);
-        terrain.setName(name);
-        PxBoxGeometry terrainGeometry = new PxBoxGeometry(0.5f, 0.5f, 0.5f);   // PxBoxGeometry uses half-sizes
 
-
-        final int minY = chunk.getWorld().getMinHeight();
-        final int maxY = chunk.getWorld().getMaxHeight();
-        for (int x = 0; x <= 15; ++x) {
-            for (int y = minY; y <= maxY; ++y) {
-                for (int z = 0; z <= 15; ++z) {
-                    Block block = chunk.getBlock(x, y, z);
-                    if (!areNeighboursEmpty(chunk.getWorld(), block) || block.getBoundingBox().getVolume() == 0)
-                        continue;
-
-                    PxShape shape = physics.createShape(terrainGeometry, defaultMaterial, true, defaultShapeFlags);
-                    shape.setSimulationFilterData(tmpFilterData);
-                    PxVec3 tmpVec2 = new PxVec3(x + 0.5f, y + 0.5f, z + 0.5f);
-                    tmpPose.setP(tmpVec2);
-                    tmpVec2.destroy();
-                    shape.setLocalPose(tmpPose);
-                    terrain.attachShape(shape);
-                    terrainShapes.add(shape);
-                }
-            }
-        }
-
+        // create a large static box with size 20x1x20 as ground
+        PxVec3 tmpVec = new PxVec3(0f, -61f, 0f);
+        tmpPose.setP(tmpVec);
+        PxBoxGeometry groundGeometry = new PxBoxGeometry(100f, 0.5f, 100f);   // PxBoxGeometry uses half-sizes
+        PxShape groundShape = physics.createShape(groundGeometry, defaultMaterial, true, defaultShapeFlags);
+        PxRigidStatic ground = physics.createRigidStatic(tmpPose);
+        groundShape.setSimulationFilterData(tmpFilterData);
+        terrainShapes.add(groundShape);
+        ground.attachShape(groundShape);
         defaultShapeFlags.destroy();
         tmpFilterData.destroy();
         tmpPose.destroy();
-        terrainGeometry.destroy();
-
-        actor = terrain;
     }
 
     /**
@@ -89,22 +65,5 @@ public class PhysxTerrain {
     public void release() {
         actor.release();
         terrainShapes.forEach(PxBase::release);
-    }
-
-    /**
-     * ブロックの隣に空洞があるか判定する
-     *
-     * @param level 世界
-     * @param pos   座標
-     * @return ブロックの隣に空洞があるかどうか
-     */
-    private static boolean areNeighboursEmpty(World level, Block pos) {
-        return pos.getY() >= level.getMaxHeight() || pos.getY() <= level.getMinHeight()
-                || (pos.getY() < level.getMaxHeight() - 1 && pos.getRelative(BlockFace.UP).isEmpty())
-                || (pos.getY() > level.getMaxHeight() && pos.getRelative(BlockFace.DOWN).isEmpty())
-                || pos.getRelative(BlockFace.NORTH).isEmpty()
-                || pos.getRelative(BlockFace.EAST).isEmpty()
-                || pos.getRelative(BlockFace.SOUTH).isEmpty()
-                || pos.getRelative(BlockFace.WEST).isEmpty();
     }
 }
